Engine_cleanup: avoid unlink(NULL) on shutdown
`Engine_cleanup()` might be called twice, the second time with `Run.httpd.socket.unix.path` already freed

src/http/engine.c:530:24: runtime error: null pointer passed as argument 1, which is declared to never be null
/usr/include/unistd.h:825:48: note: nonnull attribute specified here

+++ src/http/engine.c
@@ -525,7 +525,7 @@ void Engine_stop() {
 
 void Engine_cleanup() {
         myServerSocketsCount = 0;
-        if (Run.httpd.flags & Httpd_Unix)
+        if (Run.httpd.flags & Httpd_Unix && Run.httpd.socket.unix.path)
                 unlink(Run.httpd.socket.unix.path);
 }
